refactor(cli): :recycle: Refactor coat logo ASCII creation
