+++ packages/cli/tsconfig.json
@@ -12,5 +12,6 @@
     "resolveJsonModule": true,
     "strict": true,
     "target": "esnext"
-  }
+  },
+  "exclude": ["build"]
 }
